+++ apps/autotest/models/TST/public/forms/tst_service-module-test-fm.js
@@ -1,0 +1,4 @@
+exports.mount = function (params) {
+  $App.dialogInfo('mounted with' + JSON.stringify(params))
+}
+