+++ packages/adminui-pub/CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [5.8.3]
+### Fixed
+ - `$App.doCommand` instantiate form with type `module` by calling `mount` function (as for vue form)
+  
 ### Changed 
  - i18 key `oshibkaVvoda` renamed to `fieldValidationError` and moved to up-pub